app/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.10.0
passlib==1.7.4
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
